OP no more
All rendering issues appear to be removed and the keys have been fixed
as best i can. their should also be less confusion of where to put the
IP address to connect to a server. Lastly a cap has been set on how many
ships there can be in a fleet so their will no longer be massive uber
fleets