No More Clunkers
Two new ship types - Screamer and Thunder - have been added in place of
the Clunker and a new weapon - Thumper - has been added to accompany a
rattler on Thunders. This update also comes with changes to the way
weapons are mounted on ships; now in addition to being able to swivel
guns are mounted with an offset from the center.

The two new ships have been designed allow for two unique forms of
combat and to greatly extend time necessary to win a battle with
increased shielding all round and an inteceptor like Screamer against a
slow heavy Thunder