The Drums!!!
Audio has been added thanks to the efforts of Dylan and there have been
a few additions to the game and balance modifications. Weapons are now
more powerful so that ship to ship combat takes less time, a few
improvements to AI functionality and a complete rewrite of the AI's
collision intelligence plus optimisation and NEW ZOOM FUNCTIONALITY
toggled with the 'Z' key works at all times. Lastly changes have been
made to the networking so that the Client will automatically disconnect
if messages are less than 10/s and there has been optimisation to that
code as well