Pilot Training
For now there are a few AI Ships that spawn in and fly in circles when
the server starts as target practice for Clients. But Multiple Clients
are able to connect to the one server and are therefore able to fight
each other.

Notes:
- Users can now use the settings menu to set their own key bindings
- The Host button will throw an error for Users due to the lack of an
installer to put files in the appropriate places as of yet
- There is the occasional bug where killing another ship will kick the
Client from the Server but they can simply reconnect
- The graphics files used by the Client are in a zip in the graphics
folder